+++ aframe/aframe-ar.js
@@ -10,7 +10,7 @@ AFRAME.registerSystem('artoolkit', {
                 },
 		detectionMode : {
 			type: 'string',
-			default: 'color_and_matrix',
+			default: 'mono',
 		},
 		matrixCodeType : {
 			type: 'string',
@@ -174,7 +174,7 @@ AFRAME.registerComponent('artoolkitmarker', {
 			this.data.patternUrl = THREEx.ArToolkitContext.baseURL+'../data/data/patt.hiro'
 		}else if( this.data.preset === 'kanji' ){
 			this.data.type = 'pattern'
-			this.data.patternUrl = THREEx.ArToolkitContext.baseURL+'../data/data/patt.kanji'			
+			this.data.patternUrl = THREEx.ArToolkitContext.baseURL+'../data/data/patt.kanji'
 		}else {
 			console.assert( this.data.preset === '', 'illegal preset value '+this.data.preset)
 		}